run.py
#!/usr/bin/env python3
"""
Time Series Forecasting Ablation Study - Main Entry Point

Vertical slice implementation for the complete pipeline.
Usage: python run.py --mode=full

Modes:
- full: Complete ablation study with all models and evaluations
- data: Data loading and preprocessing only
- models: Model training and evaluation only
- ensemble: Ensemble creation and evaluation only
"""

import argparse
import yaml
import logging
from pathlib import Path
from datetime import datetime

# Import modules (will be implemented)
# from src.data_loader import DataLoader
# from src.preprocessor import DataPreprocessor
# from src.feature_selector import FeatureSelector
# from src.models import *
# from src.validator import DataValidator
# from src.ensemble import EnsembleManager
# from src.evaluator import ModelEvaluator
# from src.utils import save_results

def setup_logging():
    """Setup logging configuration"""
    logging.basicConfig(
        level=logging.INFO,
        format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
        handlers=[
            logging.FileHandler('pipeline.log'),
            logging.StreamHandler()
        ]
    )

def load_config():
    """Load configuration from YAML files"""
    config = {}
    config_dir = Path('config')
    
    for config_file in ['horizons.yaml', 'models.yaml', 'paths.yaml']:
        config_path = config_dir / config_file
        if config_path.exists():
            with open(config_path, 'r') as f:
                config[config_file.replace('.yaml', '')] = yaml.safe_load(f)
        else:
            logging.warning(f"Config file not found: {config_path}")
    
    return config

def run_full_pipeline(config):
    """Complete ablation study pipeline"""
    logging.info("Starting full ablation study pipeline...")
    
    # Phase 1: Data Loading
    logging.info("Phase 1: Data Loading")
    # data_loader = DataLoader(config['paths'])
    # train_data, test_data = data_loader.load_all_data()
    
    # Phase 2: Data Preprocessing
    logging.info("Phase 2: Data Preprocessing")
    # preprocessor = DataPreprocessor(config)
    # train_clean, test_clean = preprocessor.process_data(train_data, test_data)
    
    # Phase 3: Feature Selection
    logging.info("Phase 3: Feature Selection")
    # feature_selector = FeatureSelector(config['feature_selection'])
    # selected_features = feature_selector.select_features(train_clean)
    
    # Phase 4: Model Training
    logging.info("Phase 4: Model Training")
    # models = train_all_models(config['models'], train_clean, selected_features)
    
    # Phase 5: Evaluation
    logging.info("Phase 5: Evaluation")
    # evaluator = ModelEvaluator(config)
    # results = evaluator.evaluate_all_models(models, test_clean)
    
    # Phase 6: Ensemble Creation
    logging.info("Phase 6: Ensemble Creation")
    # ensemble_manager = EnsembleManager(config['ensemble'])
    # ensemble_results = ensemble_manager.create_ensembles(models, results)
    
    # Phase 7: Results Saving
    logging.info("Phase 7: Results Saving")
    # save_results(results, ensemble_results, config['paths'])
    
    logging.info("Pipeline completed successfully!")

def run_data_only(config):
    """Data loading and preprocessing only"""
    logging.info("Running data pipeline only...")
    # Implementation for data-only mode

def run_models_only(config):
    """Model training and evaluation only"""
    logging.info("Running models pipeline only...")
    # Implementation for models-only mode

def run_ensemble_only(config):
    """Ensemble creation and evaluation only"""
    logging.info("Running ensemble pipeline only...")
    # Implementation for ensemble-only mode

def main():
    """Main entry point"""
    parser = argparse.ArgumentParser(description='Time Series Forecasting Ablation Study')
    parser.add_argument('--mode', type=str, default='full',
                       choices=['full', 'data', 'models', 'ensemble'],
                       help='Pipeline mode to run')
    
    args = parser.parse_args()
    
    # Setup
    setup_logging()
    logging.info(f"Starting pipeline in mode: {args.mode}")
    
    # Load configuration
    config = load_config()
    logging.info("Configuration loaded successfully")
    
    # Run appropriate pipeline
    try:
        if args.mode == 'full':
            run_full_pipeline(config)
        elif args.mode == 'data':
            run_data_only(config)
        elif args.mode == 'models':
            run_models_only(config)
        elif args.mode == 'ensemble':
            run_ensemble_only(config)
        
        logging.info("Pipeline completed successfully!")
        
    except Exception as e:
        logging.error(f"Pipeline failed: {str(e)}")
        raise

if __name__ == "__main__":
    main()
