requirements.txt
# Core dependencies for time series forecasting ablation study
# Based on successful notebook implementations

# Data manipulation
numpy>=1.24.0
pandas>=2.0.0
polars>=0.20.0

# Machine Learning
lightgbm>=4.0.0
xgboost>=2.0.0
catboost>=1.2.0
scikit-learn>=1.3.0

# Deep Learning (BNN)
tensorflow>=2.13.0
tensorflow-probability>=0.21.0

# Feature Selection and Interpretability
shap>=0.42.0

# Configuration and Utilities
pyyaml>=6.0
tqdm>=4.65.0
pathlib

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Time Series (optional)
statsmodels>=0.14.0

# Development and Testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Jupyter (for notebooks)
jupyter>=1.0.0
ipywidgets>=8.0.0

# Logging and monitoring
loguru>=0.7.0

# Parallel processing
joblib>=1.3.0

# Data validation
great-expectations>=0.18.0

# Performance monitoring
memory-profiler>=0.61.0
psutil>=5.9.0
